pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# --- PROJECT METADATA & DEPENDENCIES ---
[project]
name = "ml_core"
version = "0.1.0"
description = "My MLOps project for SURF / UvA"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "h5py",
    "matplotlib",
    "seaborn",
    "pandas",
    "torch",
    "torchvision",
    "torcheval",
    "tensorboard",
    "umap-learn",
    "tqdm",
    "pyyaml",
    "python-dotenv",
    # Dev tools
    "pytest",
    "ruff",
    "pre-commit",
]

# --- TOOL CONFIGURATION ---
[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [".git", "__pycache__", "venv"]
# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
# E: pycodestyle errors
# F: Pyflakes
# I: isort (Import Sorting - MLOps critical!)
# B: flake8-bugbear (Likely bugs)
select = ["E", "F", "I", "B"]
# B905: no strict=True needed for zip
ignore = ["E501", "B905", "E741"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
